Ability to not load plugin due to its uncompatibility
l
